Support additional timedeltas
